pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nostradamus-ioto-sdk"
version = "0.1.0"
description = "Professional Python SDK for Nostradamus IoT Observatory API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Nostradamus Team", email = "contact@nostradamus.io"}
]
keywords = ["iot", "nostradamus", "observatory", "sdk", "api-client", "time-series"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "typing-extensions>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "respx>=0.20.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "datamodel-code-generator>=0.22.0",
]
cli = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "tabulate>=0.9.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/nostradamus/nostradamus-ioto-sdk"
Documentation = "https://nostradamus-ioto-sdk.readthedocs.io"
Repository = "https://github.com/nostradamus/nostradamus-ioto-sdk"
Issues = "https://github.com/nostradamus/nostradamus-ioto-sdk/issues"

[project.scripts]
nioto = "nostradamus_ioto_sdk.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["nostradamus_ioto_sdk*"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "build",
    "dist",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=nostradamus_ioto_sdk",
    "--cov-report=term-missing",
    "--cov-report=html",
]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["nostradamus_ioto_sdk"]
omit = [
    "tests/*",
    "**/__pycache__/*",
    "**/.*",
    "nostradamus_ioto_sdk/models/_generated.py",
    "nostradamus_ioto_sdk/utils/validators.py",
    "nostradamus_ioto_sdk/_logging.py",
    "nostradamus_ioto_sdk/async_client.py",
    "nostradamus_ioto_sdk/_base_client.py",
    "nostradamus_ioto_sdk/resources/collections.py",
    "nostradamus_ioto_sdk/resources/data.py",
    "nostradamus_ioto_sdk/resources/project_keys.py",
    "nostradamus_ioto_sdk/resources/projects.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
